Move offers to enterprise-user-subsidy
Offers are no longer used only in the dashboard sidebar, so this is a more appropriate location
