+++ src/components/dashboard/sidebar/offers/index.js
@@ -1,3 +1,1 @@
-export { fetchOffers } from './data/actions';
-export { default as reducer } from './data/reducer';
 export { default as Offer } from './Offer';
