+++ src/components/enterprise-user-subsidy/offers/index.js
@@ -1,0 +1,2 @@
+export { fetchOffers } from './data/actions';
+export { default as reducer } from './data/reducer';
